+++ CHANGELOG.md
@@ -5,6 +5,8 @@ project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## unreleased
 
+## v3.9.0
+
 Feature:
 - Add support for `value()` method on actors (#152)